+++ src/pages/map.jsx
@@ -2,13 +2,12 @@ import { MapContainer, TileLayer, Marker, Popup, ScaleControl, useMap } from 're
 import L from 'leaflet';
 import { useEffect, useState, useMemo, useCallback } from 'react';
 import 'leaflet/dist/leaflet.css';
-import { MenuBar } from '../components/menuBar';
 import { Header } from '../components/utils/header';
 import { Footer } from '../components/utils/footer';
 import { useParams } from "react-router-dom";
-import { WindSpeedLayer, TemperatureLayer, CloudLayer, RainViewerData, HybridLayer } from './layers';
-import { WeatherPopupContent } from './utils/weatherVariables';
-import { CustomZoomControl, CustomAttributionControl, MapMode } from './utils/mapElements';
+import { WindSpeedLayer, TemperatureLayer, CloudLayer, RainViewerData, HybridLayer } from '../components/layers';
+import { WeatherPopupContent } from '../components/utils/weatherVariables';
+import { CustomZoomControl, CustomAttributionControl, MapMode, MenuBar } from '../components/utils/mapElements';
 import { useDeviceDetect } from '../hooks/useDeviceDetect';
 
 // Custom component to style popups based on map mode
@@ -187,10 +186,10 @@ export default function ShowMap(props) {
                         zIndex={1}
                         attribution='&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors | &copy; <a href="https://carto.com/attributions">CARTO</a>'
                         url={(mapType === 'light') ?
-                            "https://{s}.basemaps.cartocdn.com/light_all/{z}/{x}/{y}{r}.png" :
-                            "https://cartodb-basemaps-{s}.global.ssl.fastly.net/dark_all/{z}/{x}/{y}.png"
+                            "https://{s}.basemaps.cartocdn.com/rastertiles/voyager_labels_under/{z}/{x}/{y}{r}.png" :
+                            "https://{s}.basemaps.cartocdn.com/rastertiles/dark_all/{z}/{x}/{y}{r}.png"
                         }
-                        subdomains={(mapType === 'light') ? "abcd" : "abcd"}
+                        subdomains="abcd"
                     />
                     <RainViewerData show={rainLayerChoice} opacity={layerOpacity} />
                     <WindSpeedLayer show={windLayerChoice} opacity={layerOpacity}/>